+++ LearningManagementSystem/LearningManagementSystem/Repositories/LessonRepository.cs
@@ -15,42 +15,49 @@ namespace LearningManagementSystem.Repositories
             _context = context;
         }
 
-        public IEnumerable<Lesson> GetLessonsByCourseId(string courseId)
-        {
-            var query = _context.Lessons
-                                .Include(l => l.Course)
-                                .AsQueryable();
-
-            if (!string.IsNullOrEmpty(courseId))
-            {
-                query = query.Where(l => l.CourseId == courseId);
-            }
-
-            return query.ToList();
-        }
-
-        public Lesson GetById(string lessonId) // Đổi từ GetLessonById thành GetById
+        public IEnumerable<Lesson> GetAll()
         {
             return _context.Lessons
-                           .Include(l => l.Course)
-                           .FirstOrDefault(l => l.LessonId == lessonId);
+                .Include(l => l.Course)
+                .ToList();
+        }
+
+        public Lesson GetById(string lessonId)
+        {
+            return _context.Lessons
+                .Include(l => l.Course)
+                .FirstOrDefault(l => l.LessonId == lessonId);
+        }
+
+        public IEnumerable<Lesson> GetLessonsByCourse(string courseId)
+        {
+            return _context.Lessons
+                .Include(l => l.Course)
+                .Where(l => l.CourseId == courseId)
+                .ToList();
         }
 
         public void Add(Lesson lesson)
         {
             _context.Lessons.Add(lesson);
-            _context.SaveChanges();
         }
 
         public void Update(Lesson lesson)
         {
             _context.Lessons.Update(lesson);
-            _context.SaveChanges();
         }
 
-        public void Delete(Lesson lesson)
+        public void Delete(string lessonId)
         {
-            _context.Lessons.Remove(lesson);
+            var lesson = _context.Lessons.FirstOrDefault(l => l.LessonId == lessonId);
+            if (lesson != null)
+            {
+                _context.Lessons.Remove(lesson);
+            }
+        }
+
+        public void Save()
+        {
             _context.SaveChanges();
         }
     }
